refactor: update store structure
