+++ store/moduels/vue-pixel/index.ts
